+++ activejob/CHANGELOG.md
@@ -1,2 +1,9 @@
+*   Deprecate `Rails.application.config.active_job.use_big_decimal_serialize`.
+
+    *Rafael Mendonça França*
+
+*   Remove deprecated primitive serializer for `BigDecimal` arguments.
+
+    *Rafael Mendonça França*
 
 Please check [7-1-stable](https://github.com/rails/rails/blob/7-1-stable/activejob/CHANGELOG.md) for previous changes.